Refactor code structure and update perk values
This commit refactors the directory structure of the special items to chestitems and general items to util. It also renames the variable 'path' to 'pathName' in the AvailablePerks enum class, and the enum's value's names have been updated to be capitalized. Moreover, the perk settings for a user have been added to the Game class and the Knife item class has been moved to the 'chestitems' folder. In the PerkSettings class, the perks' descriptions, costs, and active statuses have been set to new specified values.