Title: Extend BasePerk to manage game perks
This commit extends BasePerk to provide methods for giving and removing perks, and renames perk classes to consistently end in "Perk". Gives each perk two new functionalities; allow players to transitively connect the perk on or off at will, from an inventory GUI that appears after someone right-clicks on a defined item in-game, and abides by the subject's click event instruction.

This update aims to make the class hierarchy more meaningful and the code clearer. It will also allow devs to easily manage and maintain new game perks by adding them to the 'AvailablePerks' enum and the 'PerkSettings' class.